refactor(cli): migrate preflight.js to TypeScript
Convert bin/lib/preflight.js (357 lines) to src/lib/preflight.ts with
full type definitions for all opts objects and return types. The old
file becomes a thin re-export shim so existing consumers are unaffected.

Changes:
- Typed interfaces: PortProbeResult, MemoryInfo, SwapResult, and all
  opts types (CheckPortOpts, GetMemoryInfoOpts, EnsureSwapOpts)
- Extract parseLsofLines helper to reduce duplication in checkPortAvailable
- Incorporate #1227 fix: sudo -> sudo -n (non-interactive) for lsof retry
- Co-locate tests: test/preflight.test.js -> src/lib/preflight.test.ts
  converted to expect-style with type narrowing
- Add real net probe tests (EADDRINUSE detection on occupied ports)
- Fix co-located test imports to go through dist/ for coverage attribution
- Add targeted dashboard and validation branch tests for ratchet

612 CLI tests pass. Coverage ratchet passes. No user-facing behavior changes.

Relates to #924 (shell consolidation). Supersedes #1227.
